feat(instituto v2): Se sigue trabajando en el ejemplo InstitutoV2

+++ 2oTrimestre/REST/Instituto.V2/src/main/java/es/iespuertodelacruz/pme/institutov2/entity/Alumno.java
@@ -23,7 +23,7 @@ public class Alumno implements Serializable {
 	@Column(length=50)
 	private String apellidos;
 
-	private BigInteger fechanacimiento;
+	private Long fechanacimiento;
 
 	@Column(length=50)
 	private String nombre;
@@ -51,11 +51,11 @@ public class Alumno implements Serializable {
 		this.apellidos = apellidos;
 	}
 
-	public BigInteger getFechanacimiento() {
+	public Long getFechanacimiento() {
 		return this.fechanacimiento;
 	}
 
-	public void setFechanacimiento(BigInteger fechanacimiento) {
+	public void setFechanacimiento(Long fechanacimiento) {
 		this.fechanacimiento = fechanacimiento;
 	}
